# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Personas y Aulas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/horarios": {
            "get": {
                "tags": [
                    "Grados"
                ],
                "summary": "Listar horarios",
                "operationId": "b7d2926fab8b10719ce5669cbd456891",
                "responses": {
                    "200": {
                        "description": "Lista de horarios"
                    }
                }
            },
            "post": {
                "tags": [
                    "Grados"
                ],
                "summary": "Crear horario",
                "operationId": "8ee72160883ee946cf29142423ecab47",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "dia",
                                    "hora_inicio",
                                    "hora_fin"
                                ],
                                "properties": {
                                    "dia": {
                                        "type": "string",
                                        "example": "Lunes"
                                    },
                                    "hora_inicio": {
                                        "type": "string",
                                        "example": "08:00"
                                    },
                                    "hora_fin": {
                                        "type": "string",
                                        "example": "10:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Horario creada"
                    }
                }
            }
        },
        "/api/horarios/{id}": {
            "put": {
                "tags": [
                    "Grados"
                ],
                "summary": "Actualizar horario",
                "operationId": "2cd1d40182ef6a1e27ef8bb4b6c6d569",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la horario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "dia",
                                    "hora_inicio",
                                    "hora_fin"
                                ],
                                "properties": {
                                    "dia": {
                                        "type": "string",
                                        "example": "Lunes"
                                    },
                                    "hora_inicio": {
                                        "type": "string",
                                        "example": "08:00"
                                    },
                                    "hora_fin": {
                                        "type": "string",
                                        "example": "10:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Horario actualizada"
                    },
                    "404": {
                        "description": "Horario no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Grados"
                ],
                "summary": "Eliminar horario",
                "operationId": "32d442c036ba3491c0a723cf6d1c9afc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la horario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Horario eliminada"
                    },
                    "404": {
                        "description": "Horario no encontrada"
                    }
                }
            }
        },
...
        "/api/periodos": {
            "get": {
                "tags": [
                    "Periodos"
                ],
                "summary": "Listar periodos",
                "operationId": "d869f9ab2fcd15a45725b779e09d71ef",
                "responses": {
                    "200": {
                        "description": "Lista de periodos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Periodos"
                ],
                "summary": "Crear periodo",
                "operationId": "9321abd6dc046107d6ed2705b7ee5442",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "fecha_inicio",
                                    "fecha_fin"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "2023-2024"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Periodo creada"
                    }
                }
            }
        },
        "/api/periodos/{id}": {
            "put": {
                "tags": [
                    "Periodos"
                ],
                "summary": "Actualizar periodo",
                "operationId": "2739500d3d30e53ca83371e40f3ddeb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la periodo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "fecha_inicio",
                                    "fecha_fin"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "2023-2024"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Periodo actualizada"
                    },
                    "404": {
                        "description": "Periodo no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Periodos"
                ],
                "summary": "Eliminar periodo",
                "operationId": "65fa2a708900a9ddd1b5a060146095ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la periodo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Periodo eliminada"
                    },
                    "404": {
                        "description": "Periodo no encontrada"
                    }
                }
            }
        },
...
    },
...
}